pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "app using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
kickoff = "app.main:kickoff"
plot = "app.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
